Move MarcLanguageCodeList back to internal_model
