+++ pipeline/transformer/transformer_calm/src/main/scala/uk/ac/wellcome/platform/transformer/calm/transformers/CalmLanguages.scala
@@ -1,7 +1,7 @@
 package uk.ac.wellcome.platform.transformer.calm.transformers
 
 import weco.catalogue.internal_model.work.{Language, LanguageNote}
-import weco.catalogue.source_model.marc.MarcLanguageCodeList
+import weco.catalogue.internal_model.marc.MarcLanguageCodeList
 
 import scala.util.matching.Regex
